requirements.txt
flask
Flask-Session
flask-bootstrap
flask-socketio
python-dotenv
requests
eventlet
